Extract TraktApi to own file
